+++ ts/packages/node/src/index.ts
@@ -10,7 +10,7 @@ export * from './stores/IniFile';
 export * from './stores/AWSParameterStore';
 export * from './stores/AWSSecretsManager';
 export * from './stores/AzureKeyVault';
-export * from './stores/etcd';
+export * from './stores/etcd3';
 export * from './stores/GCPSecretManager';
 export * from './stores/HashiCorpVault';
 export * from './stores/KubernetesSecret';
